move ValidatingStore to extra package
