+++ core/src/jsMain/kotlin/dev/fritz2/binding/store.kt
@@ -228,98 +228,9 @@ open class RootStore<D>(
 }
 
 /**
- * A [ValidatingStore] is a [RootStore] which also contains a [Validation] for its model and by default applies it
- * to every update.
- *
- * This store is intentionally configured to validate the data on each update, that is why the [validateAfterUpdate]
- * parameter is set to `true` by default.
- *
- * There might be special situations where it is reasonable to disable this behaviour by setting [validateAfterUpdate]
- * to `false` and to prefer applying the validation individually within custom handlers, for example if a model should
- * only be validated after the user has completed his input or if metadata is needed for the validation process.
- * Then be aware of the fact, that the call of the [validate] function actually updates the [messages] [Flow] already.
- *
- * If the new data is not passed to the store's state after validating it, the messages are probably out of sync with
- * the actual store's state!
- * This could lead to false assumptions and might produce hard to detect bugs in your application.
- *
- * @param initialData first current value of this [Store]
- * @param validation [Validation] function to use at the data on this [Store].
- * @param validateAfterUpdate flag to decide if a new value gets automatically validated after setting it to the [Store].
- * @param id id of this [Store]. Ids of [SubStore]s will be concatenated.
- */
-open class ValidatingStore<D, T, M>(
-    initialData: D,
-    private val validation: Validation<D, T, M>,
-    val validateAfterUpdate: Boolean = true,
-    override val id: String = Id.next()
-) : RootStore<D>(initialData, id) {
-
-    private val validationMessages: MutableStateFlow<List<M>> = MutableStateFlow(emptyList())
-
-    /**
-     * [Flow] of the [List] of validation-messages.
-     * Use this [Flow] to render out the validation-messages and to detect the valid state of the current [data] [Flow].
-     */
-    val messages: Flow<List<M>> = validationMessages.asStateFlow()
-
-    /**
-     * Resets the validation result.
-     *
-     * Beware that cleaning the messages should not be done, if the [data] [Flow] remains in an invalid state.
-     * Please refer to the class's description for details about the need for a sound data and messages state.
-     *
-     * @param messages list of messages to reset to. Default is an empty list.
-     */
-    protected fun resetMessages(messages: List<M> = emptyList()) {
-        validationMessages.value = messages
-    }
-
-    /**
-     * Validates the given [data] by using the optional [metadata] to update the
-     * [messages] list and returning them.
-     * Use this method from inside your [Handler]s to publish
-     * the new state of the validation result via the [messages] flow.
-     *
-     * @param data data to validate
-     * @param metadata optional metadata for validation
-     * @return [List] of messages
-     */
-    protected fun validate(data: D, metadata: T? = null): List<M> =
-        validation(data, metadata).also { validationMessages.value = it }
-
-    init {
-        if(validateAfterUpdate) this.syncBy(this.handle<D> { _, newValue ->
-            validate(newValue)
-            newValue
-        })
-    }
-}
-
-/**
- * Checks if a [Flow] of a [List] of [ValidationMessage]s is valid.
- */
-val <M : ValidationMessage> Flow<List<M>>.valid: Flow<Boolean>
-    get() = this.map { it.valid }
-
-/**
  * Convenience function to create a simple [RootStore] without any handlers, etc.
  *
  * @param initialData first current value of this [Store]
  * @param id id of this store. Ids of [SubStore]s will be concatenated.
  */
 fun <D> storeOf(initialData: D, id: String = Id.next()) = RootStore(initialData, id)
-
-/**
- * Convenience function to create a simple [ValidatingStore] without any handlers, etc.
- * The created [Store] validates its model after every update automatically.
- *
- * @param initialData first current value of this [Store]
- * @param validation [Validation] instance to use at the data on this [Store].
- * @param id id of this [Store]. Ids of [SubStore]s will be concatenated.
- */
-fun <D, T, M> storeOf(
-    initialData: D,
-    validation: Validation<D, T, M>,
-    id: String = Id.next()
-) = ValidatingStore(initialData, validation, true, id)
